Move helper methods from scripts to utils.py
